requirements.txt
django
pillow
flake8
django-bootstrap-v5
djangorestframework
django-filter